Update ex26_1.c
This file demonstrates taking user inputted strings, and placing them contiguously in memory using a single pointer of char type. Strings are separated by a null byte, and the array location of each string is represented by an int array, idx. To print out the strings, it is necessary to use a nested for loop; the outer for loop increments the location in idx (therefore pointing to different strings), while the inner loop increments the location in to_search (moving up characters until it reaches a null byte).